requirements.txt
apscheduler==3.10.4
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
uiautomator2==3.4.0
uiautodev==0.10.0
adb_shell==0.4.4
adbutils==2.9.3
requests==2.32.5
retry==0.9.2
retry2==0.9.5
Pillow==11.3.0
pyusb==1.3.1
lxml==6.0.1
python-dotenv==1.0.1
colorlog==6.8.2
